update linker tests to have descriptive names
Update linker table driven tests to have descriptive names and run with
t.Run (for easier targeting/debugging of a single test). Enable the
'paralleltest' linter which will ensure that all tests can run in
parallel and we properly reassign variables inside a loop so they run
properly in parallel.