adds ast.Sequence and ast.Item (#48)
The only functional change is that the AST no longer keeps trailing
newlines inside line comments' raw text. Since these are expected
in source code info, they are added during a later step (in the
`sourceinfo` package).

This adds a new generic API, ast.Sequence, for navigating through
a file's elements. It supports both ast.Token, which are leaf nodes
in the AST (and do not include comments), and ast.Item, which is
a new type that represents any lexical element, which includes
both tokens and comments.