handful of small API tweaks (#76)
I discovered some issues in the API when trying to port protoparse
(buf's old compiler) to actually use protocompile under its hood.

* Allow `parser.Result` to be used as search result. This is more
efficient that requiring the caller to supply either an AST or a
descriptor proto. The protoparse package has a weird feature (that I
intentionally chose to omit from protocompile) to try to infer paths.
But that means we have to actually parse the files and all of their
dependencies, and then potentially rewrite file names, _before_ linking.
This allows that step to supply the `parser.Result`, after potentially
modifying the descriptor proto's file name, and allows protocompile to
use the result directly (vs. reconstructing the proto from the AST or
using a proto without an AST).
* Fix up usage of `os.IsNotExist`. The `os` package [Go
docs](https://pkg.go.dev/os#IsNotExist) state that newer code should
instead use `errors.Is(err, fs.ErrPermission)`.
* Fix `linker.files` receivers in all methods. This is not a
particularly big deal, but it's a large enough struct that it doesn't
make much sense to copy around the whole thing on the stack.