improve errors when string literal contains bad escapes (#78)
Previously, the lexer would not consume an entire string literal if it encountered
invalid escape sequences therein. That would make parser recovery generate a
lot of confusing error messages for the content after the failure point.

Now we consume the entire literal, even if there are multiple errors therein, and
can report them all. This makes the output much easier to understand and better
matches the output of `protoc` (which handles tokenization this way already).