Speed up pathological fuzz test case (#86)
Fuzz tests are actually configured for jhump/protoreflect. But as of
recent change, that repo's protoparse package now uses this
protocompile repo as an implementation detail.

The fuzz test complains that a pathological input causes a timeout,
which could be an abuse/DOS vector. After profiling it, I discovered
that 73% of the time was being spent in computing node info, all in
calls to NodeInfo.Start(). But the info wasn't actually used: it was
eagerly computed before checking for error conditions. Now it's
no longer eagery computed.

Furthermore, the input has an incredibly deeply nested structure
of messages. So this also eliminates some other wasted work by
not even constructing the descriptor hierarchy that is arbitrarily
deep, stopping when the depth limit is reached. Previously, it
would create the entire hierarchy and then do a depth-first
traversal to decide if the nesting depth is too great.