Extension resolution in message literals (#87)
Recently, I noticed a discrepancy between protocompile and protoc. I can't
remember why I was playing with something like this, but I discovered that
the C++ text format does not resolve references in quite the same way as
the parser. That leads to a strange inconsistency between resolving
extensions in option _names_ vs. those that appear inside message literals
in option _values_. This PR (sadly) reproduces that inconsistency here
so that protocompile matches protoc.