Fix another fuzz test timeout (#90)
The fuzz test in question constructed a pathological input case whose
performance is unexpectedly terrible. The issue is actually with huge,
but invalid string literals. The issue was actually introduced recently
with #78. The way that was resolved was to accumulate _all escape
errors_ inside the string literal and then report them at the end. It turns
out, this is needlessly expensive if, for example, there are 1000s of
errors inside a string literal, but the configured error reporter only cares
about the first (e.g. it fails fast). At that point, the cost of constructing the
1000s of errors and computing their source position info is wasteful and
is what the profiler found to be the issue in this test case.

So now, instead of accumulating all of them and then reporting them at
the end, we report them as we go, and simply skip reporting (and skip
creation of error and computation of source position) if the handler is
done (e.g. it returns a non-nil error when handling one of the escape
errors).