Update protoc to v22.0 (#100)
This repo uses `protoc` for verification -- to make sure that things
like options interpretation and source code info match what the
reference compiler produces.

This updates the version of `protoc` used to the latest release v22.0.
This uncovered discrepancies: changes/fixes in `protoc` that required
changes to protocompile code to match. This includes a very minor
change to the canonical binary format for options and more significant
improvements to comment attribution in source code info.