Support interpreting options without an AST (#110)
The interpretation of options currently relies on the AST, mainly to
provide better source positions in error messages, including pinpointing
issues inside of message literals (which `protoc` cannot do since it
doesn't create an AST that models message literals).

However, the downside is that the logic fails with unexpected `nil`
values when asked to interpret options in an unlinked descriptor (with
no AST). This addresses that gap. A small number of functions are now
forked between the path with an AST vs. the path without. The biggest
difference between the two paths is for message literals -- without an
AST, we must use the text format to parse the aggregate values that are
stored in uninterpreted options in descriptors, and must also handle
them differently when it comes to the canonical bytes.