Improve grammar to eliminate all conflicts (#138)
There were previously a number of shift/reduce conflicts in the
grammar, all due to ambiguity. Luckily, `goyacc`'s conflict resolution
strategy seemed correct as there are no known defects due to the
ambiguities. But this clean-up is valuable to ensure that there are no
strange, yet-to-be-discovered bugs from the ambiguity.