Preliminary editions work (#165)
This is just preliminary work. It adds support to the parser, and that's
about all. The only other thing it adds is code to prevent usage of
"features" in proto2 or proto3.

NOTE: This does not actually enable editions support!
The new functionality is behind an internal flag that is only enabled
of tests. We will remove the flag and enable support for editions
outside of tests once the implementation is complete.