Validate and resolve features when compiling source files that use editions (#169)
This touches what is already the most complicated part of
compilation: interpreting options. And the features resolution is not
particularly simple since it must work with dynamic versions of the
`FeatureSet` message, in case a user supplies their own version of
`google/protobuf/descriptor.proto`.

The checks in here, validating the `FeatureSet` definition, all mirror
checks done in `protoc`.

The job of resolution is to actually evaluate _all_ features and stamp
them into every element. Many features can be defined at the file level,
which means they are then "inherited" by all definitions in the file. So
when we resolve features for a particular element, we then need to treat
that resolved set as "defaults" for all children of that element.
Resolution is also where we validate that a feature is only applied to a
valid element (based on the feature's `target_type` option).

Further additional complications: now that standard options include one
that is a nested message (`features`), that tickled more issues where
the code unconditionally used dynamic messages, which later cause issues
if we try to use reflection to set a field on a non-dynamic message.
This situation results in a panic because the setter wants the generated
message type, not a dynamic message. So we have to handle that when
processing message literals in source.

This also updates the parser validation tests to compare the results to
`protoc`.

**NOTE:** Editions functionality is still disabled for actual consumers
of the library. It is only enabled for tests.