Update to match Editions behavior of latest protoc (v25.0)  (#191)
Between v24 and v25, the implementation or editions inside of `protoc`
changed quite rather drastically. Most of the complexity having to do
with computation of feature defaults and feature set resolution has been
completely removed 🎉.

This still does not perform some of the checks for field features that
`protoc` performs, so that still needs to come in a subsequent branch.
But this does bring this repo up to the same level of
`protoc`-compatibility it had before, but this time with `protoc` v25
(instead of v24). It also does add one add'l check: that the `packed`
field option is not used with editions.