Support linking with dynamic resolution
This commit adds linker.LinkDynamic, which supports lookups
at runtime and does not depend on a predefined File slice.

A new func was added to stay source compatible,
otherwise a call with []linker.File would break the api.

The new func drops the *Symbols, as they are
rarely previously known in this use case.

To support reuse of symbols, a type assertion was added to
compile.go that uses the relevant function for linking.