Add function for stripping source retention options from a descriptor (#250)
This adds a helper function to the `options` sub-package that can strip
"source retention" options from a descriptor. These are options that
should only be retained in the descriptor in source form -- like when
manipulated by a compiler or code generator -- and should not be present
at runtime. Stripping these options results in a descriptor that could
safely be embedded in generated code.