Support extension declarations (#263)
This adds validation of the extension declaration configuration and then
also adds enforcement for extension ranges that have declarations.

The validation and checks are all based on the same logic in `protoc`
([here](https://github.com/protocolbuffers/protobuf/blob/main/src/google/protobuf/descriptor.cc#L8031-L8035)
and
[here](https://github.com/protocolbuffers/protobuf/blob/main/src/google/protobuf/descriptor.cc#L7092-L7095)).

The one area where we go above and beyond `protoc` is that
`protocompile` will check that an extension is only used once by a
declaration _globally_ (as best it can anyways); `protoc` on the other
hand only checks that an extension is only used once by a declaration
within a single message. So it allows the same fully-qualified extension
to be referenced in declarations from different messages. This doesn't
make sense because a single extension can only extend one message -- so
if it's a valid extension for one message, it _can't_ be a valid
extension for a different one.