Move internal editions code to new internal/editions package (#280)
This is a pre-factoring step so we can expose some editions-related
helper functions from the `protoutil` package without ending up with an
import cycle. So this moves the shared stuff out of the `linker` package
and into an internal one. It also consolidates the editions-related
stuff into one place, so it moves some things from `internal` ->
`internal/editions`.