Linker tests allow compiler to report multiple errors; fix ensuing issues with redundant/inaccurate errors reported (#281)
This is another pre-factoring step. Currently, when interpreting an
option value, the value is a "leaf" in the error checking. That means it
returns either a value or an error. If it returns an error, the caller
will handle it. If the configured error handler returns nil, the
operation keeps going, to allow collecting multiple errors in a single
pass instead of failing fast. But I want to change this so that we can
report multiple errors even about a single option value. The main use
for this is to report all relevant errors inside a message literal,
which is a complex composite value that could contain more than one
problem. (Currently, we have to fail fast since it's at a higher level in
the flow where we actually handle the error using the configured
handler.)

So this updates the linker tests to report _all_ errors, to exercise
the code path that will be changing. And, in doing so, I revealed
numerous existing problems. In some cases, it would follow up
with inaccurate/confusing error messages. And in some other
cases, it would report duplicate/redundant errors. So this fixes
those issues as well. The most complicated thing in here is an
improvement to reporting locations in the file for an option
message where a required field was not set.