Fix flakiness of non-deterministic test case (#282)
In #281, a flakiness was introduced for a test case that allows
multiple possible outcomes due to concurrency -- when
an import cycle exists, more than one goroutine could be processing
multiple files participating in the cycle. As soon as one goroutine
discovers the cycle, it aborts the compile. But it's possible for
another goroutine to also report the cycle it discovers (the same cycle,
but starting from a different file) before it receives the cancellation
signal.

So this makes the test lenient: instead of requiring only one of the
possible errors to be reported, it accepts both being reported as well.