Add editions helper functions for resolving features to protoutil (#283)
These helpers, in particular `protoutil.ResolveFeature` and
`protoutil.ResolveCustomFeature`, will be used from updated checks for
`buf breaking`, which will allow the tool to understand the
features-related semantics of the schema.

This way, it can correctly report issues with incompatible changes to
features in editions source files. And it can also allow changing a
file's syntax (like migrating from proto2 or proto3 to editions) as long
as there are no actual semantic changes to the schema.