Disallow options that use message-set-wire-format if Protobuf runtime can't support them (#284)
The protobuf-go runtime currently only supports message set wire format
when built with the `protolegacy` build tag. Most users don't use this
tag, but protocompile allows sources to be processed that define
messages with this legacy feature.

The problem arises if a custom option refers to such a message: in that
case, the resulting descriptors returned by protocompile will not be
serializable. Since the protobuf-go runtime doesn't support message set
wire format, it will serialize the options as if they were normal wire
format messages, which will be incorrect.

So this module now checks if the protobuf-go dependency supports
message set wire format (by trying to create a very simple file
descriptor that uses it). If it does not, it will report an error when
interpreting options if any option tries to use message set wire format.