Add lookup methods to linker.Symbols; improve performance/reduce allocations (#286)
This adds a lookup methods, which can be used to improve the efficiency
of a resolver that is based on a `linker.Files`: instead of O(n) to
query each file for an item, we can look the item up from the symbol
table and directly determine the file it is in.

To further aid efficiency, this adds some benchmarks and then optimizes
the `Symbols.getPackage` function so both symbol table construction and
lookup are more efficient. In particular, instead of `strings.Split`,
which must allocate a slice, we now use a new `nameEnumerator` type,
which can be used to iterate through name components without allocation.