Last change for Editions: enforce feature lifetimes, and make it available to users (#301)
This updates to protobuf-go v1.34.1, which provides the latest
`descriptor.proto` updates in its `descriptorpb` package. With that, we
are able to implement the final change before allowing end-users to use
protocompile with editions: enforcing feature lifetimes.

So this PR implements that last change and also makes Editions
generally available: it's no longer disabled-by-default, requiring
opt-in via "testing" method.

The last change is a check to verify that a feature isn't being used outside
its expected lifetime. This PR also adds an exported helper to the root
protocompile package: IsEditionSupported. This advertises to consumers
of the package which editions it can compile.