Add wellknownimports package, for providing actual source code for the standard imports (#310)
This adds wellknownimports.WithStandardImports as an alternative
to protocompile.WithStandardImports. The alternative provides
source code for the imports instead of just providing the embedded
descriptors from generated Go code. This allows users to make use
of the extension declarations added to `descriptor.proto` in v27.0
(they are absent from embedded descriptors since they are marked
as source-only retention).