Fix panic that can occur when interpreting options in lenient mode (#331)
The panic fix is tiny. But this commit is bigger because other
changes/improvements were called for: most of the code changes are to
improve error handling when in lenient mode. After I fixed the panic,
the test case was failing in a different way, due to an issue with how
errors (even in lenient mode) were still being passed to an error
reporter and causing the stage to fail.

This issue was introduced in #279. That PR moved things around, pushing
the responsibility of calling `interp.reporter.HandleError` down, so
that interpreting a single option could potentially report multiple
errors (instead of failing fast and only reporting the first).

But when in lenient mode, we don't actually want to send those errors to
the reporter: sending the error to the reporter meant the error would
get memoized and then returned in subsequent handle calls, which could
cause the process to fail when it should be lenient and also can cause
it to report the wrong error in lenient mode. So now we demarcate the
parts of the process where errors are tolerated (i.e. where lenience
mode is actually activated) with a new field on the interpreter that is
examined when errors are reported.