Add a new diagnostics library (#347)
This is the first PR of many to implement a new frontend for
protocompile. New packages will live under `protocompile/experimental`
until we're ready to promote the next protocompile verison.

This change adds:
- `internal/golden`, a golden tests framework.
- ~~`internal/width`, a port of Rust's `unicode-width` for calculation
the approximate terminal display width of multilingual strings.~~ Using
`uniseg` for this, works great!
- `report.proto`, a Protobuf for diagnostics.
- `experimental/report`, the package itself.
- `experimental/report/testdata`, a suite of tests for the report
renderer.

The fussiest part of this PR is `experimental/report/renderer.go`. I
have tried to include tests for all of the corner cases, and I have
attempted to document it so that which mysterious layout trick is
written down, lest I forget.

There's one intentionally broken test: it's very difficulty to support
users who want to use RTL scripts in diagnostics, because not only does
that cause the text direction to flip, it flips around things that are
not part of the RTL script. Bidi handling is not worth it, and not even
Rust tries to get this right.