Allow queries to check if their dependencies were recomputed (#372)
This change adds `incremental.Result.Changed`, which is set if the
result was not cached from a prior call to `incremental.Run`. This is
*not* the same as "this call to `Resolve` recomputed the query", because
the value of `Changed` is the same for all calls to `Resolve` of that
query (i.e., it is deterministic).

This is intended for cases where we want to merge the result of a huge
number of queries in one query, and want to implement our own "partial
caching" of that query. For example, consider a query which merges the
symbol tables of many files. If only one of the files has changed, it is
much cheaper to modify only the symbols belonging to that file, than to
re-merge the whole table. This enables that optimization.

This PR includes a test that hammers the executor to ensure that we
always get consistent results for Changed.