Implement a new lexer (and split `Token` and related code out of `experimental/ast`) (#358)
This PR implements a new Protobuf lexer under `experimental/parser`. It
is a fairly lenient lexer that parses many things `protoc` does not, and
diagnoses them.

Along with this I have separated the token parts of `experimental/ast`
into its own package. This was actually a really good idea, and makes
the token code (arguably the fussiest part of the AST library) easier to
follow. It's a good sign when three files become eight.

I also did some work in `experimental/report` to make it so there is
Only One True Span Type, `report.Span`. This side-steps a lot of the
nastiness around carrying line and column information around; it is now
truly only computed on-demand when diagnostics are rendered. Also,
hooray for the massive `experimental/report` test suite.