Merge `report.File` and `report.IndexedFile` (#373)
There are basically no uses of `report.File` on its own anymore, so this
PR merges it with `report.IndexedFile` to be used everywhere. This
simplifies callers somewhat, too.

This also adds `report.File.Span()`, a shorthand for creating spans in a
file, and `report.File.EOF()`, which returns the EOF span.
`IndexedFile.Search` is now called `File.Location`.