Add a canonical query for requesting the contents of a source file (#371)
I settled on adding a simple `source.Opener` interface that provides the
necessary method, a `queries.File` query whose key is the opener plus a
path, and I included types that can be used for building an Opener that
are careful about errors.

The new `queries` package will contain other Protocompile-specific
queries, such as `queries.AST`.

This change also adds `incremental.Executor.EvictWithCleanup`. This is
identical to `Evict` except that it allows callers to run cleanup code
inside of the eviction critical section. This allows eviction to be
synchronized with the mutation of other structures. For example, suppose
there is a `source.Map` that we want to modify an entry of. This
function enables us to ensure that calls to `Run` do not see a stale
value in the map.

```go
unlock := exec.EvictAndCleanup([]any{queries.File{Opener: myMap, path: myPath})}, func() {
  (*myMap)[myPath] = newContent // Write happens while the eviction lock is held.
})
```