Add package internal/intern (#379)
This PR adds a simple interning table with 32-bit IDs. It currently
implements interning using a slice plus a reverse index, protected by a
read-write lock. This will be used so that qualified Protobuf names can
be represented as slices: `name []intern.ID`.

The table supports up to two billion interned names, which is unlikely
to ever be overflown by a single complier session. Also, ASCII
identifiers up to five characters (which may contain, but not end with,
a period) are encoded directly in the ID using the LLVM bitcode `char6`
encoding. This removes the need to hit the reverse index for very small
common identifiers, such as "data", "value", "error", "bytes", and
"msg".

It may be worthwhile to investigate in the future whether using 64-bit
IDs (or some `[n]uint64`), which would allow identifiers with at most
`64 / 6 ~ 10` characters to be encoded inline. Whether this is worth
doing will depend on what most common identifier lengths on the BSR are.