Add internal/taxa, which contains canonical names of various syntactic entities (#383)
This package is used throughout the parser both to render diagnostics
and as a cheap way of tracking containing contexts (such as, for
example, when legalizing what defs a message is allowed to contain).

The package includes `taxa.Noun`, an enum that will likely be expanded
over time. `taxa.Classify` classifies an actual syntactic element using
a large type switch. `taxa.Place` is used to provide relative location
information, and `taxa.Set` is an efficient set of `taxa.Noun` values.