Refactor various "standard library extension" helpers to live together (#391)
This PR moves the `iters` and `slicesx` packages into `internal/ext`,
since I expect that over time this will grow (particularly since we seem
to be adding lots of iterator adapters).

I've also taken the liberty to move all of the sketchy `unsafe` tricks
throughout the compiler (except for `pathkey.go`, since it's so
specialized) into an `unsafex` package. This way we can at least have
comments explaining what each of these tricks does, so they can be
reused as needed.

I've also added `slicesx.Get` and `slicesx.Last`, which are ports of
functions that Rust has on its slices. There's a lot of places where
tedious (and easy to mess up) bounds checks that I plan to replace with
calls to these.