Use unsafex.StringAlias in linker.pathKey (#400)
This fixes the old reflection-based approach to convert a slice to an
array (so it is comparable and can be used as a map key) to instead use
`unsafex.StringAlias`. This elides some extra copying that was
happening in the old implementation and is also hopefully simpler to
read and understand than the old way (though it does mean that keys in
the map will be less scrutable in a debugger since the underlying slice
data will make for completely unreadable/junk strings).