Add "error nodes" for types and expressions (#398)
Types and expressions can form complex parse trees, but sometimes
something goes wrong in a sub-part of the tree. For example, a dict
field might have an invalid expression as its value. Rather than use
nil, which typically means "missing" and has no span information, we can
use `ExprError` with a span containing the skipped tokens.

This also means that we can have parser functions like e.g.
`tryParseExpr()` which can signal nil when the tokens that follow cannot
begin an expression, but `ExprError` upon committing to a parse.