Refactor how diagnostics are pushed through `incremental`; add `queries.AST` (#404)
The `NonFatal` functionality of the incremental package has proven to be
a poor abstraction. What I essentially wanted was a way to have queries
generate diagnostics, which I could then deduplicate and collect in the
end.

Collecting diagnostics from dependencies is incorrect, because A -> B, A
-> C, B -> D, C -> D would mean diagnostics for D contain the
diagnostics from A twice. The rough idea was to instead stash reports in
the `NonFatal` area, which required the somewhat unnatural-feeling
`report.AsError` type. But this is unnecessary ceremony: all that
`NonFatal` will ever be used for is for stashing reports, so the
incremental framework should Just Do That. It's not a general library,
after all.

This PR replaces `Task.NonFatal` with `Task.Report`, which is a report
included with each task. When `Run` completes, it collects the set of
all queries that were computed (possibly from cache) and merges their
reports, and sorts it to eliminate non-determinism.

It is not immediately clear to me if having tasks return `(v T, fatal
error)` is still useful. Perhaps `(v T, ok bool)` may be more
appropriate, since that error should be logged as a diagnostic and will
probably just get thrown away.