Make comma-list parsing really robust (#403)
This PR adds more diagnostics for comma-delimited list of types and
expressions, and ensures that commas are optional pretty much everywhere
for generating a valid AST, while at the same time diagnosing all
missing and extraneous commas. This ensures a match with the behavior
advertised in #393.

Includes a test that exercises various corner cases.

This PR also adds `Span.RuneRange` and `Span.Rune`, to aid in
constructing suggestion spans.