Implement machine-applicable suggestions in diagnostics (#392)
This PR adds `report.SuggestEdits`, which allows diagnostics to suggest
changes to an annotated span. This will be used for surfacing
quick-fixes in the LSP (for example, for suggesting deleting unused
imports, adding a `syntax` declaration, adding missing field numbers,
and so on).

The compiler is also able to render these as a simple diff, similar to
some of rustc's diagnostics. I've tried to keep the data model simple:
no diffing actually happens, the diagnostics are expected to generate
the diffs manually. All the renderer does is convert the edits into
hunks to show to the user. This PR contains examples in the lexer that
shows how it is meant to be used.