Add keyword information to `token.Token` (#454)
This change adds [keyword.Keyword], which is an enum of all the special
"grammar particles" the compiler recognizes, which includes punctuation.
Comparing against these is simpler and less error-prone than than
comparing against a string such as `"message"`.

This PR updates the parser to use keyword comparisons instead of string
comparisons.

In order to fit the keyword information into tree tokens, we are forced
to burn three bits of the token offset. The offset is still 26 bits
wide, which puts the maximum distance between two braces at around 32
million.

One annoying thing is the significant overlap between `Predeclared`,
`Keyword`, and `taxa.Noun`. it might be nice to add some mechanism to
`internal/enum` to allow importing another enum as a subset... but that
feels like more work than I particularly want to put into this.