Drop go 1.22 support (#460)
We have made the decision to only maintain support for the two most
recent versions of Go. Thus, we can drop 1.22, which allows us to drop
many polyfills related to iterators.

This PR does not update all of the iterator calls to use for loops. That
will happen in a series of self-contained followups.