Implement the legalizer (#438)
This PR adds the legalizer: a collection of diagnostics that follow
parsing to eliminate and diagnose invalid constructs that the parser
accepted in order the make progress.

This PR also adds a new `syntax` package for e.g. `syntax.Proto2` and
other editions knowledge. It adds functions for classifying
`predeclared.Name`s and some new `taxa.Noun`s for naming concepts in
diagnostics relevant to the legalizer. Parsing is also tweaked in a few
places to generate better diagnostics.

---------

Co-authored-by: Doria Keung <doriable@users.noreply.github.com>